move a type
