+++ note-api/src/gitApi.ts
@@ -1,0 +1,1 @@
+//
