+++ types/git.ts
@@ -1,4 +1,15 @@
+import { noteIndex } from './../types'
 export type gitRepoDetails = {
 	owner: string
 	repo: string
+}
+
+export interface IGitApi {
+	saveFile: (
+		content: string,
+		fullFilePath: string,
+		commitMessage: string
+	) => Promise<{ commitSha: string }>
+	getFile: (filePath: string) => Promise<{ content }>
+	getFiles: (path: string | undefined, extension: 'md') => Promise<noteIndex>
 }
